User API Created

+++ blogging/src/main/java/com/bloggingApi/blogging/dao/UserDao.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bloggingApi.blogging.modal.User;
 
-public interface UserDao extends JpaRepository<User, Long> {
+public interface UserDao extends JpaRepository<User, Integer> {
 
+	
 }
